Fixing Broken auth when using UBP_SITEURL
The original code removed all url parts but host and path in the UBP_SITEURL conditional in the get_siteurl() method. I've added a property called url_parts that gets assigned parsed url values before the reconstruction of the url. This way the url_parts can be used in the get_auth method without the auth being removed form the UBP_SITEURL